Issue #324: Move order services into separate folder
